requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
jinja2==3.1.4
torch==2.6.0
torchvision==0.21.0
numpy==2.1.1
pillow==10.4.0
python-multipart==0.0.9
opencv-python==4.10.0.84

